empty string validation added

+++ src/test/java/org_test/example/CalculateTDDTest.java
@@ -8,19 +8,19 @@ public class CalculateTDDTest {
     @Test
     public void test1() {
         CalculateTDD calculateTDDTest=new CalculateTDD();
-        int result=calculateTDDTest.add("");
+        Long result=calculateTDDTest.add("");
         assertEquals(0,result);
     }
     @Test
     public void test2() {
         CalculateTDD calculateTDDTest=new CalculateTDD();
-        int result=calculateTDDTest.add("1");
+        Long result=calculateTDDTest.add("1");
         assertEquals(1,result);
     }
     @Test
     public void test3() {
         CalculateTDD calculateTDDTest=new CalculateTDD();
-        int result=calculateTDDTest.add("1,5");
+        Long result=calculateTDDTest.add("1,5");
         assertEquals(6,result);
     }
 }
